:truck: Moved ProfileButton
Moved the profile button from /lib/widgets/home/profile to /lib/widgets/home/search.
